% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CorrTab.R
\name{CorrTab}
\alias{CorrTab}
\title{Takes a dataset with all numeric vectors and calculates the correlation between each pair}
\usage{
CorrTab(data)
}
\arguments{
\item{data}{a dataset of numeric columns}
}
\value{
a table of correlation values
}
\description{
Takes a dataset with all numeric vectors and calculates the correlation between each pair
}
\examples{
mock_data <- data.frame(pred1 = rep(1:4,10),pred2=rep(1:5,8),missvar=as.numeric(c("NA",rep(1:2,19),"NA")))
CorrTab(mock_data)
}
